UsbNet: separate tinyusb callbacks to dedicated file
